requirements.txt
Babel==0.9.6
Jinja2==2.6
Markdown==2.2.1
MarkupSafe==0.15
Paste==1.7.5.1
PasteDeploy==1.5.0
Pygments==1.6
SQLAlchemy==0.8.0b2
Tempita==0.5.1
WebError==0.10.3
WebHelpers==1.3
WebOb==1.2.3
paster==0.8
postmarkup==1.2.0
psycopg2==2.4.6
pyOpenSSL==0.13
pyramid==1.4
pyramid-debugtoolbar==1.0.4
pyramid-jinja2==1.6
pyramid-tm==0.7
pyramid-webassets==0.7.1
python-dateutil==2.1
pytz==2012j
raven==3.1.11
repoze.lru==0.6
six==1.2.0
textile==2.1.5
times==0.6.1
transaction==1.4.0
translationstring==1.1
w3lib==1.2
webassets==0.8
wsgiref==0.1.2
zope.component==4.0.2
zope.deprecation==4.0.2
zope.event==4.0.2
zope.interface==4.0.3
zope.sqlalchemy==0.7.1
